Rename deprecated predictors db file
